refactor: Seperate swiper to component module
