+++ index.html
@@ -41,18 +41,13 @@
 						</div>
 					</div>
 				</nav>
+
+				<i class="fa-solid fa-chevron-left swiper"></i>
 				<div class="press__grid-wrap active">
-					<div class="press__grid-wrap__button">
-						<i class="fa-solid fa-chevron-left swiper"></i>
-					</div>
 					<div class="press__grid-wrap__grid"></div>
-					<div class="press__grid-wrap__button">
-						<i class="fa-solid fa-chevron-right swiper"></i>
-					</div>
 				</div>
 				<div class="press__list-wrap">
 					<ul class="press__list__nav" role="tablist"></ul>
-
 					<div class="press__list__news" role="tabpanel">
 						<div class="press__list__news-wrap">
 							<div class="press__list__news-top"></div>
@@ -63,6 +58,7 @@
 						</div>
 					</div>
 				</div>
+				<i class="fa-solid fa-chevron-right swiper"></i>
 			</section>
 		</div>
 	</body>
