+++ src/css/view/swiper.css
@@ -13,9 +13,9 @@
 }
 
 .swiper.fa-chevron-left {
-	left: -10%;
+	left: -7%;
 }
 
 .swiper.fa-chevron-right {
-	right: -10%;
+	right: -7%;
 }
